+++ CHANGELOG.md
@@ -6,12 +6,15 @@ file.
 ## [Unreleased]
 ### Added
 - Now trace into executed binaries
-- Added `--avoid-cfg-tarpaulin` flag to remove `--cfg=tarpaulin` from the `RUSTFLAGS`
+- Added `--avoid-cfg-tarpaulin` flag to remove `--cfg=tarpaulin` from the
+`RUSTFLAGS`
 
 ### Changed
 - Address offset mapping has been added which allows us to compile binaries
 without changing the relocation model
 - Tie match patterns to a single logical line
+- Check if unable to read file to string and skip source analysis for it if 
+that's the case
 
 ### Removed